Patient Data Entry tool: add readme (JIRA AEKR-1368)

+++ eCRF_OncoCologne/install_ecrf_oncocologne.sh
@@ -4,6 +4,11 @@ set -e
 export POSTGRES_PASSWORD=gyZcrz9Zzp@8waX8
 export OPAL_SUPER_USER_PASSWORD=9gM6W8#qjUX8E7B8
 
+fresh_install=false
+if [[ $(docker volume inspect postgres_data | grep -q 'Error') != 0 ]]; then
+  fresh_install=true
+fi
+
 echo "pull docker images"
 docker pull harbor.honeur.org/ecrf/postgres:0.2
 docker pull harbor.honeur.org/ecrf/app:0.2
@@ -21,10 +26,14 @@ docker run -d --name honeur_ecrf_postgres --network honeur-net --volume postgres
 sleep 5s
 
 echo "create eCRF app container"
-docker run -d --name honeur_ecrf_app --network honeur-net --volume static_volume:/code/entrytool/assets --env OPAL_SUPER_USER_PASSWORD=$OPAL_SUPER_USER_PASSWORD --env OPAL_DB_USER=postgres --env OPAL_DB_PASSWORD=$POSTGRES_PASSWORD --env OPAL_DB_NAME=postgres --env OPAL_DB_HOST=honeur_ecrf_postgres --env OPAL_DB_PORT=5432 --env OPAL_FLUSH_DB=true --env OPAL_ENABLE_USER_DB=false --env OPAL_ENABLE_LDAP=false harbor.honeur.org/ecrf/app:0.2 gunicorn -b 0.0.0.0:8000 entrytool.wsgi
-sleep 15s
-docker stop honeur_ecrf_app
-docker rm honeur_ecrf_app
+if $fresh_install; then
+  echo "Fresh install"
+  docker run -d --name honeur_ecrf_app --network honeur-net --volume static_volume:/code/entrytool/assets --env OPAL_SUPER_USER_PASSWORD=$OPAL_SUPER_USER_PASSWORD --env OPAL_DB_USER=postgres --env OPAL_DB_PASSWORD=$POSTGRES_PASSWORD --env OPAL_DB_NAME=postgres --env OPAL_DB_HOST=honeur_ecrf_postgres --env OPAL_DB_PORT=5432 --env OPAL_FLUSH_DB=true --env OPAL_ENABLE_USER_DB=false --env OPAL_ENABLE_LDAP=false harbor.honeur.org/ecrf/app:0.2 gunicorn -b 0.0.0.0:8000 entrytool.wsgi
+  sleep 15s
+  docker stop honeur_ecrf_app
+  docker rm honeur_ecrf_app
+fi
+
 docker run -d --name honeur_ecrf_app --network honeur-net --volume static_volume:/code/entrytool/assets --env OPAL_SUPER_USER_PASSWORD=$OPAL_SUPER_USER_PASSWORD --env OPAL_DB_USER=postgres --env OPAL_DB_PASSWORD=$POSTGRES_PASSWORD --env OPAL_DB_NAME=postgres --env OPAL_DB_HOST=honeur_ecrf_postgres --env OPAL_DB_PORT=5432 --env OPAL_FLUSH_DB=false --env OPAL_ENABLE_USER_DB=false --env OPAL_ENABLE_LDAP=false --restart=always harbor.honeur.org/ecrf/app:0.2 gunicorn -b 0.0.0.0:8000 entrytool.wsgi
 sleep 15s
